re-submitting ISBN validation PR
